Refactored signals
Wrapped all signal objects inside global static functions so that they're local static objects instead of global static objects.  This solves a (future) issue with the Debug class.  The reason for this move is because global static objects have an undefined order of initialization so it's possible that they might not have been initialized at certain critical points.

Signed-off-by: Geoff Shannon <earthlingzephyr@gmail.com>